remove task directory from this PR
