--- forecasting/src/autogluon/forecasting/task/forecasting/__init__.py
@@ -1,3 +1,0 @@
-from .forecasting import Forecasting
-from .predictor_legacy import ForecastingPredictorV1
-
